# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.AssetCreate:
    properties:
      amount:
        type: number
      asset_market:
        type: string
      asset_type:
        enum:
        - crypto
        - stock
        - exchange
        - commodity
        type: string
      from_asset:
        type: string
      price:
        type: number
      to_asset:
        type: string
      type:
        enum:
        - sell
        - buy
        type: string
    required:
    - amount
    - asset_type
    - from_asset
    - price
    - to_asset
    - type
    type: object
...
  requests.AssetUpdate:
    properties:
      amount:
        type: number
      id:
        type: string
      price:
        type: number
      type:
        type: string
    required:
    - id
    type: object
...
  requests.ChangeCurrency:
    properties:
      currency:
        type: string
    required:
    - currency
    type: object
  requests.ChangeFCMToken:
    properties:
      fcm_token:
        type: string
    required:
    - fcm_token
    type: object
  requests.ChangeMembership:
    properties:
      is_lifetime_premium:
        type: boolean
      is_premium:
        type: boolean
    type: object
  requests.ChangeNotification:
    properties:
      app_notification:
        type: boolean
      mail_notification:
        type: boolean
    required:
    - app_notification
    - mail_notification
    type: object
  requests.ChangePassword:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  requests.ForgotPassword:
    properties:
      email_address:
        type: string
    required:
    - email_address
    type: object
...
  requests.Register:
    properties:
      currency:
        type: string
      email_address:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - currency
    - email_address
    - password
    type: object
...
  responses.Asset:
    properties:
      asset_market:
        type: string
      asset_type:
        type: string
      current_total_value:
        type: number
      from_asset:
        type: string
      name:
        type: string
      p/l:
        type: number
      pl_percentage:
        type: number
      remaining_amount:
        type: number
      to_asset:
        type: string
      total_bought:
        type: number
      total_sold:
        type: number
    type: object
  responses.AssetAndStats:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.Asset'
        type: array
      stats:
        $ref: '#/definitions/responses.AssetStats'
    type: object
...
  responses.AssetStats:
    properties:
      commodity_assets:
        type: number
      commodity_p/l:
        type: number
      commodity_percentage:
        type: number
      crypto_assets:
        type: number
      crypto_p/l:
        type: number
      crypto_percentage:
        type: number
      currency:
        type: string
      exchange_assets:
        type: number
      exchange_p/l:
        type: number
      exchange_percentage:
        type: number
      stock_assets:
        type: number
      stock_p/l:
        type: number
      stock_percentage:
        type: number
      total_assets:
        type: number
      total_bought:
        type: number
      total_p/l:
        type: number
      total_pl_percentage:
        type: number
      total_sold:
        type: number
    type: object
...
  responses.UserInfo:
    properties:
      app_notification:
        type: boolean
      currency:
        type: string
      email_address:
        type: string
      fcm_token:
        type: string
      investing_limit:
        type: string
      is_lifetime_premium:
        type: boolean
      is_oauth:
        type: boolean
      is_premium:
        type: boolean
      subscription_limit:
        type: string
    type: object
host: rocky-reaches-65250.herokuapp.com
info:
  contact:
    email: mrntlu@gmail.com
    name: Burak Fidan
  description: REST Api of Kantan.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://rocky-reaches-65250.herokuapp.com/terms/
  title: Kantan Investment Manager API
  version: "1.0"
paths:
  /asset:
    delete:
      consumes:
      - application/json
      description: Deletes all assets by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete All Assets by User ID
      tags:
      - asset
    get:
      consumes:
      - application/json
      description: Returns assets and stats by user id
      parameters:
      - in: query
        name: assetType
        type: string
      - enum:
        - name
        - percentage
        - amount
        - profit
        in: query
        name: sort
        required: true
        type: string
      - enum:
        - 1
        - -1
        in: query
        name: type
        required: true
        type: integer
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.AssetAndStats'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Assets & Stats by User ID
      tags:
      - asset
    post:
      consumes:
      - application/json
      description: Creates asset
      parameters:
      - description: Asset Create
        in: body
        name: assetcreate
        required: true
        schema:
          $ref: '#/definitions/requests.AssetCreate'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Asset
      tags:
      - asset
    put:
      consumes:
      - application/json
      description: Updates asset log by id
      parameters:
      - description: Asset Update
        in: body
        name: assetupdate
        required: true
        schema:
          $ref: '#/definitions/requests.AssetUpdate'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: Unauthorized update
          schema:
            type: string
        "404":
          description: Couldn't find user
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Asset Log by AssetID
      tags:
      - asset
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Allows users to register
      parameters:
      - description: User registration info
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/requests.Register'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: User Registration
      tags:
      - auth
...
  /user:
    delete:
      consumes:
      - application/json
      description: Deletes everything related to user
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Deletes user information
      tags:
      - user
  /user/change-currency:
    put:
      consumes:
      - application/json
      description: Users can change their default currency
      parameters:
      - description: Set currency
        in: body
        name: changecurrency
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeCurrency'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Change User Currency
      tags:
      - user
  /user/change-notification:
    put:
      consumes:
      - application/json
      description: Users can change their notification preference
      parameters:
      - description: Set notification
        in: body
        name: changenotification
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeNotification'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Change User Notification Preference
      tags:
      - user
  /user/change-password:
    put:
      consumes:
      - application/json
      description: Users can change their password
      parameters:
      - description: Set new password
        in: body
        name: ChangePassword
        required: true
        schema:
          $ref: '#/definitions/requests.ChangePassword'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Change User Password
      tags:
      - user
  /user/forgot-password:
    post:
      consumes:
      - application/json
      description: Users can change their password when they forgot
      parameters:
      - description: User's email
        in: body
        name: ForgotPassword
        required: true
        schema:
          $ref: '#/definitions/requests.ForgotPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Couldn't find any user
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Will be used when user forgot password
      tags:
      - user
  /user/info:
    get:
      consumes:
      - application/json
      description: Returns users membership & investing/subscription limits
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Info
          schema:
            $ref: '#/definitions/responses.UserInfo'
      security:
      - ApiKeyAuth: []
      summary: User membership info
      tags:
      - user
  /user/membership:
    put:
      consumes:
      - application/json
      description: User membership status will be updated depending on subscription
        status
      parameters:
      - description: Set Membership
        in: body
        name: changemembership
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeMembership'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Change User Membership
      tags:
      - user
  /user/update-token:
    put:
      consumes:
      - application/json
      description: Depending on logged in device fcm token will be updated
      parameters:
      - description: Set token
        in: body
        name: changefcmtoken
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeFCMToken'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates FCM User Token
      tags:
      - user
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"